Add a first simple application test
At the moment, the only unit test we have covers a CSV helper. I would like to be able to write unit tests for any new features or fixes. Create a simple application test as an example of how to do such tests.

Since we are mocking the MigrationService, this isn't actually testing very much. But it should be something we can build on for more meaningful tests.

The trick here is that mocking the MigrationService avoids the need for any of the complex configuration that makes it so difficult to run the app locally.