Add `isBuyer` when adding organisation roles
This is a workaround for a bug in either SSO or its swagger spec. Previously, we were not including `isBuyer` when adding organisation roles. This is legal according to the swagger spec. I would expect that SSO would not change `isBuyer` if DM omitted it. The problem is that SSO was setting `isBuyer` to false on the basis of our requests.

This should be fixed in SSO or its swagger spec. However, given the time pressures, let's add a workaround to always provide `isBuyer`. We may be able to revert this once SSO is fixed, but this should unblock testing in the interim.

Add a unit test to check this is working.