Document how to regenerate the CII and SSO clients
The clients were originally autogenerated using swagger-codegen. I can find no documentation as to how that was done. The SSO API has changed in the interim. I could update it manually. However, the right thing to do is to work out how to get autogeneration working again. This will ensure that we can keep the client in this codebase up to date with the swagger spec without any risk of introducing bugs by making the changes manually.

Note that the DM models themselves were also generated using swagger-codegen. However, it's not included here because there's no swagger spec for it in this repo.