Reinstate old title behaviour
In 5abb1b3a6383bb752b186a163015f834450498ae I had to move the title parsing logic out of the autogenerated code. However, I made a mistake - I didn't match the previous behaviour. Previously, if the title wasn't an integer, it would be set to null. Reinstate this behaviour.

It's not the right behaviour, but we'll fix that shortly.