Add support for adding user contact details
SSO now requires the contact name, so provide it if we have it. This should allow us to start setting user contact details successfully. Add unit tests to check the SSO request is being constructed properly.

Future commits will handle partial contact data and validating the presence of the name.