Create script to run non-production database migrations
After a lot of fiddling, I've worked out how to get liquibase to do what we want. Unfortunately, this requires disabling SSL. It seems as if the SSL key the database is using a root certificate my copy of liquibase doesn't recognise. This isn't good. However, it's still useful for non-production environments in the interim.