Log at INFO level in preprod and prod
We've configured debug-level logging for sandbox and integration - https://github.com/Crown-Commercial-Service/ccs-conclave-data-migration/pull/272. We don't want debug logging in prod due to the risk of leaking sensitive data to the logs. And pre-prod should match prod.

So set an INFO log level in these higher environments. This should strike the appropriate balance of being sufficiently informative but not too spammy.