Remove timestamp precision from db changelog
Postgres doesn't support the listed precision - only 0 to 6 are valid values. I don't know how we successfully applied it with 29 digits of precision.

The postgres docs say "By default, there is no explicit bound on precision", so we're safe to remove it. The difficult part will be applying this to the environments. We need to edit the changeset because otherwise it doesn't apply locally. I think we're just going to have to force it in by clearing the checksums.